# workdir/docs/swagger.yaml
basePath: /
definitions:
  auth.Role:
    enum:
    - ADMIN
    - FORM_CREATOR
    - BASIC
    - BASIC
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleFormCreator
    - RoleBasic
    - DefaultRole
  errors.Code:
    enum:
    - invalid_argument
    - unauthorized
    - forbidden
    - not_found
    - conflict
    - internal
    type: string
    x-enum-varnames:
    - CodeInvalidArgument
    - CodeUnauthorized
    - CodeForbidden
    - CodeNotFound
    - CodeConflict
    - CodeInternal
  errors.HTTPErrorResponse:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/errors.Code'
        example: not_found
      message:
        example: form not found
        type: string
    type: object
...
  handlers.SetRoleRequest:
    properties:
      role:
        allOf:
        - $ref: '#/definitions/auth.Role'
        enum:
        - ADMIN
        - FORM_CREATOR
        - BASIC
        example: FORM_CREATOR
    type: object
  handlers.SignUpRequest:
    properties:
      password:
        description: Password is 8 to 72 bytes long.
        example: correct horse battery
        type: string
      username:
        description: 'Username is case-insensitive: 3 to 50 letters, digits, ''.'',
          ''_'' or ''-''.'
        example: alice
        type: string
    type: object
...
  handlers.TokenResponse:
    properties:
      access_token:
        description: 'AccessToken is a JWT; send it as "Authorization: Bearer <token>".'
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_at:
        example: "2026-01-01T01:00:00Z"
        type: string
      token_type:
        example: Bearer
        type: string
      user:
        $ref: '#/definitions/handlers.UserResponse'
    type: object
  handlers.UserListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handlers.UserResponse'
        type: array
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
    type: object
  handlers.UserResponse:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      role:
        allOf:
        - $ref: '#/definitions/auth.Role'
        enum:
        - ADMIN
        - FORM_CREATOR
        - BASIC
        example: BASIC
      username:
        example: alice
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'Username and password of a user: "Basic " + btoa(username + ":" +
    password).'
  title: App API
  version: "1.0"
paths:
  /auth/me:
    get:
      description: Returns the user authenticated by the bearer token or Basic credentials.
      produces:
      - application/json
      responses:
        "200":
          description: Signed-in user
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "401":
          description: Not signed in
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
      security:
      - BearerAuth: []
      - BasicAuth: []
      summary: Get the current user
      tags:
      - auth
  /auth/signin:
    post:
      description: 'Checks the HTTP Basic credentials ("Authorization: Basic " + btoa(username
        + ":" + password)) and returns a JWT access token to send as "Authorization:
        Bearer <token>".'
      produces:
      - application/json
      responses:
        "200":
          description: Signed in
          schema:
            $ref: '#/definitions/handlers.TokenResponse'
        "401":
          description: Missing or invalid credentials
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
      security:
      - BasicAuth: []
      summary: Sign in
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Creates an account with the BASIC role. Usernames are case-insensitive.
      parameters:
      - description: Credentials of the new account
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Account created
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Invalid username or password
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "409":
          description: Username already taken
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
      summary: Sign up
      tags:
      - auth
...
  /users:
    get:
      description: Lists every user by ID. Requires the ADMIN role.
      parameters:
      - default: 0
        description: Number of users to skip
        in: query
        minimum: 0
        name: offset
        type: integer
      - default: 20
        description: Maximum number of users to return
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Page of users
          schema:
            $ref: '#/definitions/handlers.UserListResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "401":
          description: Not signed in
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "403":
          description: Not an admin
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
      security:
      - BearerAuth: []
      - BasicAuth: []
      summary: List users
      tags:
      - users
  /users/{id}/role:
    put:
      consumes:
      - application/json
      description: Sets the role of another user. Requires the ADMIN role; admins
        cannot change their own role.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: New role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.SetRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "401":
          description: Not signed in
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "403":
          description: Not an admin
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
      security:
      - BearerAuth: []
      - BasicAuth: []
      summary: Change the role of a user
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    description: API key of the tenant, returned when the tenant is created.
    in: header
    name: X-API-Key
    type: apiKey
  BasicAuth:
    type: basic
  BearerAuth:
    description: JWT returned by POST /auth/signin, sent as "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"